feat(api): add centralized API fetch wrapper with auth and 401 handling
- Add src/lib/api.ts with apiFetch<T> wrapper
- Prepend VITE_API_URL to all requests
- Attach Authorization: Bearer token from useAuthStore when available
- Handle 401 by calling clearAuth() and throwing a consistent ApiError
- Refactor api-client.ts to use the new wrapper

Closes #37
