+++ src/lib/api-client.ts
@@ -1,40 +1,11 @@
 import type { Guide, Tip } from '../types/education';
+import { apiFetch } from './api';
 
-const API_BASE = import.meta.env.VITE_API_BASE_URL ?? '';
-
-export class ApiError extends Error {
-    status?: number;
-    constructor(message: string, status?: number) {
-        super(message);
-        this.status = status;
-        this.name = 'ApiError';
-    }
-}
-
-async function fetchApi<T>(endpoint: string, options?: RequestInit): Promise<T> {
-    try {
-        const response = await fetch(`${API_BASE}${endpoint}`, {
-            ...options,
-            headers: {
-                'Content-Type': 'application/json',
-                ...options?.headers,
-            },
-        });
-
-        if (!response.ok) {
-            throw new ApiError(`Failed to fetch ${endpoint}`, response.status);
-        }
-
-        return await response.json();
-    } catch (error) {
-        if (error instanceof ApiError) throw error;
-        throw new ApiError(error instanceof Error ? error.message : 'An unexpected error occurred');
-    }
-}
+export { ApiError } from './api';
 
 export const educationApi = {
-    getGuides: () => fetchApi<Guide[]>('/api/education/guides'),
-    getTip: () => fetchApi<Tip | null>('/api/education/tip'),
+    getGuides: () => apiFetch<Guide[]>('/api/education/guides'),
+    getTip: () => apiFetch<Tip | null>('/api/education/tip'),
 };
 
 export interface Round {
@@ -47,7 +18,7 @@ export interface Round {
 }
 
 export const roundsApi = {
-    getActive: () => fetchApi<Round | null>('/api/rounds/active'),
+    getActive: () => apiFetch<Round | null>('/api/rounds/active'),
 };
 
 export interface UserPrediction {
@@ -86,7 +57,7 @@ function normalizeUserPredictions(response: UserPredictionsResponse): UserPredic
 
 export const predictionsApi = {
     getUserHistory: async (userId: string) => {
-        const response = await fetchApi<UserPredictionsResponse>(`/api/predictions/user/${encodeURIComponent(userId)}`);
+        const response = await apiFetch<UserPredictionsResponse>(`/api/predictions/user/${encodeURIComponent(userId)}`);
         return normalizeUserPredictions(response);
     },
 };
